fix role and exceptions (#4)

+++ src/main/java/com/rest_api/fs14backend/author/AuthorService.java
@@ -1,35 +1,51 @@
 package com.rest_api.fs14backend.author;
 
+import com.rest_api.fs14backend.exceptions.NotFoundException;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 
 import java.util.List;
+import java.util.Optional;
 import java.util.UUID;
 
 @Service
 public class AuthorService {
-    @Autowired
-    private AuthorRepository authorRepository;
+  @Autowired
+  private AuthorRepository authorRepository;
 
-    public List<Author> findAll(){
-        return authorRepository.findAll();
+  public List<Author> findAll() {
+    return authorRepository.findAll();
+  }
+
+  public Optional<Author> findById(UUID id) {
+    Optional<Author> authorOptional = authorRepository.findById(id);
+    if (authorOptional.isEmpty()) {
+      throw new NotFoundException("Author not found");
+    }
+    return authorRepository.findById(id);
+  }
+
+  public void deleteById(UUID id) {
+    Optional<Author> authorOptional = authorRepository.findById(id);
+    if (authorOptional.isEmpty()) {
+      throw new NotFoundException("Author not found");
+    }
+    authorRepository.deleteById(id);
+  }
+
+  public Author createOne(Author author) {
+    return authorRepository.save(author);
+  }
+
+  public Author updateOne(Author newAuthor, UUID id) {
+    Optional<Author> authorOptional = authorRepository.findById(id);
+    if (authorOptional.isEmpty()) {
+      throw new NotFoundException("Author not found");
     }
 
-    public Author findById(UUID id){
-        return authorRepository.findById(id).orElse(null);
-    }
-
-    public void deleteById(UUID id) {
-        authorRepository.deleteById(id);
-    }
-
-    public Author createOne(Author author){
-        return authorRepository.save(author);
-    }
-
-    public Author updateOne(Author author, UUID id) {
-        author.setId(id);
-
-        return authorRepository.save(author);
-    }
+    return authorRepository.findById(id).map(author -> {
+      author.setName(newAuthor.getName());
+      return authorRepository.save(author);
+    }).orElseGet(() -> authorRepository.save(newAuthor));
+  }
 }
