+++ src/main/java/com/rest_api/fs14backend/book/BookService.java
@@ -1,9 +1,11 @@
 package com.rest_api.fs14backend.book;
 
+import com.rest_api.fs14backend.exceptions.NotFoundException;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 
 import java.util.List;
+import java.util.Optional;
 import java.util.UUID;
 
 @Service
@@ -16,11 +18,19 @@ public class BookService {
     return bookRepository.findAll();
   }
 
-  public Book findById(UUID id) {
-    return bookRepository.findById(id).orElse(null);
+  public Optional<Book> findById(UUID id) {
+    Optional<Book> bookOptional = bookRepository.findById(id);
+    if (bookOptional.isEmpty()) {
+      throw new NotFoundException("Book not found");
+    }
+    return bookRepository.findById(id);
   }
 
   public void deleteById(UUID id) {
+    Optional<Book> bookOptional = bookRepository.findById(id);
+    if (bookOptional.isEmpty()) {
+      throw new NotFoundException("Book not found");
+    }
     bookRepository.deleteById(id);
   }
 
@@ -28,8 +38,21 @@ public class BookService {
     return bookRepository.save(book);
   }
 
-  public Book updateOne(Book book, UUID id) {
-    book.setId(id);
-    return bookRepository.save(book);
+  public Book updateOne(Book newBook, UUID id) {
+    Optional<Book> bookOptional = bookRepository.findById(id);
+    if (bookOptional.isEmpty()) {
+      throw new NotFoundException("Book not found");
+    }
+
+    return bookRepository.findById(id).map(book -> {
+      book.setIsbn(newBook.getIsbn());
+      book.setTitle(newBook.getTitle());
+      book.setDescription(newBook.getDescription());
+      book.setPublisher(newBook.getPublisher());
+      book.setStatus(newBook.getStatus());
+      book.setPublishedDate(newBook.getPublishedDate());
+      book.setQuantity(newBook.getQuantity());
+      return bookRepository.save(book);
+    }).orElseGet(() -> bookRepository.save(newBook));
   }
 }
